requirements.txt
apache-airflow==2.9.0
apache-airflow-providers-common-sql
apache-airflow-providers-postgres
psycopg2-binary==2.9.9
google-auth==2.40.2
google-cloud-bigquery==3.17.2
pandas==2.2.3
db-dtypes==1.1.1
pyarrow==15.0.2
